requirements.txt
chromadb 
sentence-transformers 
pymupdf 
open-clip-torch 
google-generativeai
flask
flask-cors
werkzeug
